e2e: add performance test
Add a test that captures timing measurements (in 'console.log()' printouts)
for cloning various large repositories. Because these tests can be
exceptionally slow, tag them with '@slow' and make sure they are not run in
the 'default' profile. Add an 'all' profile that runs the performance tests
alongside the "regular" tests.

Signed-off-by: Victoria Dye <vdye@github.com>