Makefile: add 'e2e-test' target
Add a target to the Makefile that installs the end-to-end test NPM
dependencies then runs the tests. By default, this will only run tests
matching the "default" profile of the tests (excluding performance tests).
The 'TEST_E2E_PERF' flag, if it has a value, will instead run the tests with
the 'all' profile (including the performance tests).

Signed-off-by: Victoria Dye <vdye@github.com>