workflows: add a workflow to run end-to-end tests
Add a dedicated 'test.yml' workflow that executes the end-to-end tests.
Unlike CI, these tests are only run automatically for pull requests
(although they can be manually triggered via workflow dispatch). By default,
the tests will run with the "default" (no perf test) profile; this can be
changed via an input variable in the workflow dispatch.

Because the end-to-end tests deal with external dependencies (namely, Git),
add an 'install-dependencies.sh' script to install and configure the
dependencies on a GitHub Actions runner.

The goal of this workflow is for it to eventually be used to run multiple
types of long-running tests, such as integration tests or installer
validation tests.

Signed-off-by: Victoria Dye <vdye@github.com>